API

ClassicalSpinMC.H5 — Type

const H5 = Union{HDF5.File,HDF5.Group}

Used to define an alias for a HDF5 File or Group.

ClassicalSpinMC.MatrixObs — Type

const MatrixObs

Alias for either BinnedMatrixObs or FullMatrixObs.

ClassicalSpinMC.VectorObs — Type

const VectorObs

Alias for either BinnedVectorObs or FullVectorObs.

ClassicalSpinMC.defaultColor — Constant

Constants for all plots

ClassicalSpinMC.defaultMarker — Constant

Constants for MC plots

ClassicalSpinMC.sphereColor — Constant

Constants for common origin plots

ClassicalSpinMC.updateParameterDict — Constant

const updateParameterDict

A dictionary used to lookup the initial parameters for spin update methods.

ClassicalSpinMC.AbstractObservables — Type

abstract type AbstractObservables

This is a struct to store Observables measured during the MonteCarlo run.

!!! note In order to measure observables other than the built-in Observables type or a subset of it, you may create a subtype of AbstractObservables such as MyObservables and define the following functions.

1.performMeasurements!(observables::MyObservables, lattice::T, energy::Float64) where {T<:Lattice}

2.performPostMeasurements!(observables::MyObservables, lattice::T, β::Float64) where {T<:Lattice}

Depending on the complexity of your type, you may also have to define

1.writeObservables!(fn:H5, obs::MyObservables)

2.readObservables

if the general function cannot handle those types.

ClassicalSpinMC.BinnedMatrixObs — Type

const BinnedMatrixObs

Alias for LogBinner storing a Matrix{Float64}.

ClassicalSpinMC.BinnedObs — Type

const BinnedObs

Alias for a LogBinner storing a Float64.

ClassicalSpinMC.BinnedVectorObs — Type

const BinnedVectorObs

Alias for a LogBinner storing a Vector{Float64}.

ClassicalSpinMC.FullMatrixObs — Type

const FullMatrixObs

Alias for FullBinner storing a Matrix{Float64}.

ClassicalSpinMC.FullObs — Type

const FullObs

Alias for a FullBinner storing a Float64.

ClassicalSpinMC.FullVectorObs — Type

const FullVectorObs

Alias for a FullBinner storing a Vector{Float64}.

ClassicalSpinMC.Lattice — Type

struct Lattice

Contains information about the current spin configuration and all energy terms such as exchange interactions.

ClassicalSpinMC.Lattice — Method

function Lattice(uc::UnitCell{D}, L::NTuple{D,Int}) where {D}

Generates a Lattice from a UnitCell with periodic boundary conditions.

ClassicalSpinMC.MonteCarlo — Type

mutable struct MonteCarlo

This is used to store all information about a Monte Carlo run.

ClassicalSpinMC.MonteCarlo — Method

function MonteCarlo(lattice::T, parameters::MonteCarloParameters,
observables::Observables) where {T<:Lattice}

Constructor that generates a MonteCarloStatistics struct automatically.

ClassicalSpinMC.MonteCarlo — Method

function MonteCarlo(lattice::T, parameters::MonteCarloParameters) where {T<:Lattice}

Constructor that generates a MonteCarloStatistics struct and built-in Observables struct automatically.

ClassicalSpinMC.MonteCarloAnnealing — Type

mutable struct MonteCarloAnnealing

This is a wrapper around MonteCarlo and is used to store the multiple Monte Carlo structs created during a simulated annealing run.

ClassicalSpinMC.MonteCarloAnnealing — Method

function MonteCarloAnnealing(mc::MonteCarlo, betas::Vector{Float64})

Constructor for generating a MonteCarloAnnealing struct used for Simulated Annealing runs.

ClassicalSpinMC.MonteCarloExchange — Type

mutable struct MonteCarloExchange

This is a wrapper around MonteCarlo and is used to store the multiple Monte Carlo structs and resources needed for distributed computing during a Replica Exchange (Parallel tempering) run.

ClassicalSpinMC.MonteCarloParameters — Type

mutable struct MonteCarloParameters

This is used to store any parameters used to configure the Monte Carlo run.

ClassicalSpinMC.MonteCarloParameters — Method

function MonteCarloParameters(
beta::Float64,
thermalizationSweeps::Int,
measurementSweeps::Int)

Constructor that uses reasonable defaults for MonteCarloParameters.

ClassicalSpinMC.MonteCarloStatistics — Type

mutable struct MonteCarloStatistics

This is used to store information regarding the statistics gathered during the Monte Carlo run.

ClassicalSpinMC.Observables — Type

mutable struct Observables

Built-in Observables type that measures simple observables and serves as a template for custom observable structs.

Mandatory fields

In order to use replica exchange (parallel tempering) the labels and replicaAcceptance fields are mandatory.

ClassicalSpinMC.Observables — Method

function Observables(lattice::T) where {T<:Lattice}

Generates an Observables object for a given Lattice object.

ClassicalSpinMC.UnitCell — Type

mutable struct UnitCell

This is used to represent a UnitCell struct. For further computation it must first be converted into a Lattice struct.

ClassicalSpinMC.UnitCell — Method

function UnitCell(primitive::Vararg{SVector{D,Float64},D})

UnitCell constructor that accepts D number of D dimensional SVectors as input to use as primitives.

ClassicalSpinMC.UnitCell — Method

function UnitCell(primitive::Vararg{NTuple{D,Float64},D})

UnitCell constructor that accepts D number of D dimensional NTuples for primitives that get converted to SVectors.

ClassicalSpinMC.addAnisotropy! — Method

function addAnisotropy!(unitcell::UnitCell{D}, func::F, funcParameters::Vector{Float64}) where {D,F<:Function)

Adds an anisotropy function to the UnitCell definition along with any parameters that need to be passed along to the anisotropy function. The default anisotropy function is identity.

ClassicalSpinMC.addBasisSite! — Method

function addBasisSite!(unitcell::UnitCell{D}, positions::Vector{SVector{D,Float64}}) where {D}

Convenience function that accepts multiple positions.

ClassicalSpinMC.addBasisSite! — Method

function addBasisSite!(unitcell::UnitCell{D}, positions::Vector{NTuple{D,Float64}}) where {D}

Convenience function that accepts multiple positions and converts them into SVector internally.

ClassicalSpinMC.addBasisSite! — Method

function addBasisSite!(unitcell::UnitCell{D}, position::SVector{D,Float64}) where {D}

Adds a new basis site to the given unitcell at the given position with no self-interaction and no magnetic field.

ClassicalSpinMC.addBasisSite! — Method

function addBasisSite!(unitcell::UnitCell{D}, position::NTuple{D,Float64}) where {D}

Convenience function that accepts position as an NTuple and converts it to an SVector internally.

ClassicalSpinMC.addInteraction! — Method

function addInteraction!(unitcell::UnitCell{D}, b1::Int, b2::Int, M::Matrix{Float64}, offset::NTuple{D,Int}=Tuple(zeros(Int, D))) where {D}

Convenience function that converts to SMatrix and creates the b1 and b2 pair automatically.

ClassicalSpinMC.addInteraction! — Method

function addInteraction!(unitcell::UnitCell{D}, b1::Int, b2::Int, M::SMatrix{3,3,Float64,9}, offset::NTuple{D,Int}=Tuple(zeros(Int, D))) where {D}

Convenience function that creates the b1 and b2 pair automatically.

ClassicalSpinMC.addInteraction! — Method

function addInteraction!(unitcell::UnitCell{D}, b1::Int, b2s::Vector{Int}, M::SMatrix{3,3,Float64,9}, offsets::Vector{NTuple{D,Int}}) where {D}

Convenience function that adds the same interaction for a vector of b2 and offset values pairwise.

ClassicalSpinMC.addInteraction! — Method

function addInteraction!(unitcell::UnitCell{D}, b::Pair{Int,Int}, M::SMatrix{3,3,Float64,9}, offset::NTuple{D,Int}=Tuple(zeros(Int, D))) where {D}

Adds an interaction between spin_1 located at basis site b_1 of the given unitcell and spin_2 at basis site b_2 in a unit cell that is offset by offset lattice vectors. The exchange energy is calculated as spin_1'M.spin_2.

Please note that local interactions that involve the same site will be automatically added as an on site interaction using setInteractionOnSite!(...).

ClassicalSpinMC.anneal! — Method

function anneal!(mc::M, betas::Vector{Float64}) where {M<:MonteCarlo}

Anneals a given MonteCarlo object through the given array of betas without creating intermediate MonteCarloAnnealing and MonteCarlo objects.

ClassicalSpinMC.calcTriangles — Method

function calcTriangles(lattice::Lattice{D,N}) where {D,N}

Returns a list of all unique triangles found in the lattice consisting of the sites comprising the triangle in ascending order.

ClassicalSpinMC.conicalUpdate — Function

function conicalUpdate(v::SVector{3,Float64}, t::Float64, rng=Random.GLOBAL_RNG)

Generates a random point on a unit sphere restricted in a cone defined as being centered around the vector v and an angle t. For t=pi this is equivalent to any other sphere point picking method.

ClassicalSpinMC.createChannels — Method

function createChannels()

Creates the RemoteChannels required for worker-worker communication used in Replica Exchange (Parallel Tempering) runs.

ClassicalSpinMC.dimension — Method

function dimension(unitcell::UnitCell{D}) where {D}

Utility function to get the dimensionality of a UnitCell struct.

ClassicalSpinMC.exchangeEnergy — Method

function exchangeEnergy(s1::SVector{3,Float64}, M::SMatrix{3,3,Float64,9}, s2::SVector{3,Float64})

Calculates the energy contribution due to exchange terms.

ClassicalSpinMC.getAnisotropy — Method

function getAnisotropy(func::F, lattice::Lattice{D,N})::Float64 where {D,N,F<:Function}

Calculates the anisotropy energy term for the full lattice.

ClassicalSpinMC.getAnisotropy — Method

function getAnisotropy(::typeof(identity), lattice::Lattice{D,N})::Float64 where {D,N}

Default method for returning zero in case of no-anisotropy for the full lattice.

ClassicalSpinMC.getAnisotropy — Method

function getAnisotropy(func::F, spin::SVector{3,Float64}, parameters::NTuple{P,Float64})::Float64 where {F<:Function,P}

Calculates the anisotropy energy term for a particular spin.

ClassicalSpinMC.getAnisotropy — Method

function getAnisotropy(::typeof(identity), spin::SVector{3,Float64}, parameters::NTuple{P,Float64})::Float64 where {P}

Default method for returning zero in case of no-anisotropy for a particular spin.

ClassicalSpinMC.getChirality — Method

function getChirality(lattice::Lattice{D,N}, siteList::Vector{Vector{Int64}}) where {D,N}

Given a siteList of triangles in the lattice, computes the chirality of the lattice, where the chirality is defined as the vector triple product of spins on a triangle.

ClassicalSpinMC.getCorrelation — Method

function getCorrelation(lattice::Lattice{D,N}) where {D,N}

Measures the two-point correlation for each pair of sites in the lattice.

ClassicalSpinMC.getEnergy — Method

function getEnergy(lattice::Lattice{D,N})::Float64 where {D,N}

Calculates the energy of the current spin configuration of the lattice based on all terms defined such as exchange, onsite, field and ansiotropy terms.

ClassicalSpinMC.getEnergyDifference — Method

function getEnergyDifference(lattice::Lattice{D,N}, site::Int, newState::SVector{3,Float64})::Float64 where {D,N}

Performant method for calculating the energy difference when updating a site into a newState due to all terms such as exchange, onsite, field and ansiotropy terms.

ClassicalSpinMC.getInteractionField — Method

function getInteractionField(lattice::Lattice{D,N}, site::Int)::SVector{3,Float64} where {D,N}

Interface function to get the magnetic field on a particular site.

ClassicalSpinMC.getInteractionMatrices — Method

function getInteractionMatrices(lattice::Lattice{D,N}, site::Int)::NTuple{N,SMatrix{3,3,Float64,9}} where {D,N}

Interface function to get the exchange interactions on a particular site.

ClassicalSpinMC.getInteractionOnsite — Method

function getInteractionOnsite(lattice::Lattice{D,N}, site::Int)::SMatrix{3,3,Float64,9} where {D,N}

Interface function to get the on-site interactions on a particular site.

ClassicalSpinMC.getInteractionSites — Method

function getInteractionSites(lattice::Lattice{D,N}, site::Int)::NTuple{N,Int} where {D,N}

Interface function to get the sites that interact with a particular site via exchange terms.

ClassicalSpinMC.getMagnetization — Method

function getMagnetization(lattice::Lattice{D,N}) where {D,N}

Measures the Magnetization of the lattice spin configuration.

ClassicalSpinMC.getMagnetizationPerSite — Method

function getMagnetizationPerSite(lattice::Lattice{D,N}) where {D,N}

Measures the Magnetization for each type of basis site in the unitcell seperately.

ClassicalSpinMC.getSitePosition — Method

function getSitePosition(lattice::Lattice{D,N}, site::Int)::SVector{D,Float64} where {D,N}

Interface function to get the position of a particular site.

ClassicalSpinMC.getSpin — Method

function getSpin(lattice::Lattice{D,N}, site::Int) where {D,N}

Interface function to get the spin at a particular site.

ClassicalSpinMC.initSpinConfiguration! — Method

function initSpinConfiguration!(mc::MonteCarlo{T}) where {T<:Lattice}

Initialize the spin configuration for a Monte Carlo struct by using the function specified in MonteCarloParameters.

ClassicalSpinMC.initSpinConfiguration! — Method

function initSpinConfiguration!(lattice::Lattice{D,N}, f::Function, rng=Random.GLOBAL_RNG) where {D,N}

Initialize the spin configuration of a lattice using the passed function.

ClassicalSpinMC.initSpinConfiguration! — Method

function initSpinConfiguration!(lattice::Lattice{D,N}, f::typeof(conicalUpdate), rng=Random.GLOBAL_RNG) where {D,N}

Initialize the spin configuration of a lattice using the conicalUpdate method.

Pre-initialize for conical updates

Since conicalUpdate forms a cone around a point on a unit sphere, it cannot be used to reliably initialize a spin configuration unless it was already initialized using a different method first. This method is provided only for consistency.

ClassicalSpinMC.loadProcessedObservables — Method

function loadProcessedObservables(fn::String)

Function that loads previously processed observables into memory. Please note that since this is a compact representation of data, all technical information about the run such as the amount of sweeps / number of runs is no longer retained.

ClassicalSpinMC.localSweep — Method

function localSweep(::Function, mc::MonteCarlo{T}, energy::Float64) where {T<:Lattice}

Performs a local sweep on the current spin configuration, using the function specified in MonteCarloParameters. This covers only the case where the update function takes the rng as it's only parameter.

ClassicalSpinMC.localSweep — Method

function localSweep(mc::MonteCarlo{T}, energy::Float64) where {T<:Lattice}

Performs a local sweep on the current spin configuration using the method specified in MonteCarloParameters. This is the top-level method.

ClassicalSpinMC.localSweep — Method

function localSweep(::typeof(conicalUpdate), mc::MonteCarlo{T}, energy::Float64) where {T<:Lattice}

Performs a local sweep on the current spin configuration using the conicalUpdate method. This is a specialization of the method and not the general case.

ClassicalSpinMC.localUpdate — Method

function localUpdate(mc::MonteCarlo{T,P}, proposalSite::Int64, newSpinState::SVector{3,Float64}) where {T<:Lattice,P<:MonteCarloParameters}

Performs a local update on the current spin configuration, using the proposed new spin state.

ClassicalSpinMC.marsagliaSphereUpdate — Function

function marsagliaSphereUpdate(rng=Random.GLOBAL_RNG)

Generates a random point on a unit sphere using the marsaglia method.

ClassicalSpinMC.microcanonicalRotation — Method

function microcanonicalRotation(lattice::Lattice{D,N}, site::Int) where {D,N}

Performs an Overrelaxation step on the lattice in a deterministic way by using a π rotation.

Assumptions on interactions and ansiotropy

This assumes that there are no anisotropy contributions and no self interactions for the lattice object. For performance reasons, these checks are performed seperately.

ClassicalSpinMC.microcanonicalRotationRandom — Method

function microcanonicalRotationRandom(lattice::Lattice{D,N}, site::Int, rng=Random.GLOBAL_RNG) where {D,N}

Performs an Overrelaxation step on the lattice in a non-deterministic way by using a random rotation.

Assumptions on interactions and anisotropy

This assumes that there are no anisotropy contributions and no self interactions for the lattice object. For performance reasons, these checks are performed seperately.

ClassicalSpinMC.microcanonicalSweep! — Method

function microcanonicalSweep!(mc::MonteCarlo{T}) where {T<:Lattice}

Performs a microcanonical (overrelaxation) sweep on current spin configuration, based on parameters from MonteCarloParameters.

ClassicalSpinMC.microcanonicalSweep! — Method

function microcanonicalSweep!(lattice::Lattice{D,N}, rounds::Int, rng=Random.GLOBAL_RNG) where {D,N}

Performs a microcanonical (overrelaxation) sweep on the current spin configuration. This is particularly useful for low-temperature simulations where the spins get stuck.

ClassicalSpinMC.performMeasurements! — Method

function performMeasurements!(observables::Observables, lattice::T, energy::Float64) where {T<:Lattice}

Performs measurements and is called during measurement sweeps.

Energy measurement

Since the current energy of the spin configuration is calculated as part of the Monte Carlo sweep, it is passed on directly.

ClassicalSpinMC.performMeasurements! — Method

function performMeasurements!(observables::Observables, lattice::T, energy::Float64) where {T<:Lattice}

Generic method to perform measurements.

For every field (F) in observables, gets the method "get(F)" from Main.

!!! note "Performance" This method is not performant since it calls a seperate method for all observables regardless of any dependence between them.

ClassicalSpinMC.performPostMeasurements! — Method

function performPostMeasurements!(observables::Observables, lattice::T, β::Float64) where {T<:Lattice}

Performs measurements and is called after all Monte Carlo sweeps are performed. It is intended to do any post measurements based on the collect observables or final state of the system.

ClassicalSpinMC.piRotation — Method

function piRotation(spin::SVector{3,Float64}, axis::SVector{3,Float64})

Special case of rotateAboutAxis for a rotation by an angle π.

Assumptions on normalization

This assumes that the vector and axis are normalized for performance reasons.

ClassicalSpinMC.printStatistics! — Method

function printStatistics!(mc::MonteCarlo{T}; replica=false) where {T<:Lattice}

Print statistics from the current status of the MonteCarlo run.

ClassicalSpinMC.processObservables! — Method

function processObservables!(fn::Vector{String}, obs::Vector{Symbol}, saveLocation::String)

Function that reads data from a vector of filenames and then saves them into a single HDF5 file. Can be used to collect a mean observable for a single β across runs or to collect an observable acrosss temperatures.

ClassicalSpinMC.processObservables! — Method

function processObservables!(fn::Vector{Vector{String}}, obs::Vector{Symbol}, saveLocation::String)

Function that reads data from a vector of vector of filenames and then saves them into a single HDF5 file.

Assumes that the files in the inner vector are a single run across temperatures and the outer vector is iterating over a collection of runs.

Please note that the collection of runs are assumed to have exactly the same β values.

ClassicalSpinMC.processObservables! — Method

function processObservables!( temperatures::Vector{Float64}, observables::Matrix{Measurement{T}}, obs::Vector{Symbol}, f::H5) where {T}

Function to save an observables matrix into a H5 object.

ClassicalSpinMC.readCheckpoint — Method

function readCheckpoint(filename::String)

Interface function to read a MonteCarlo checkpoint.

Checkpoints

Checkpoints use Serialization and thus are not recommended for long term storage as they may not be supported on a different Julia version.

ClassicalSpinMC.readLattice — Method

function readLattice(fn::H5)

Reads the Lattice from a H5 object.

ClassicalSpinMC.readMonteCarlo — Method

function readMonteCarlo(filename::String)

Reads the MonteCarlo object from a HDF5 file.

ClassicalSpinMC.readMonteCarlo — Method

function readMonteCarlo(fn::H5)

Reads the MonteCarlo object from a H5 object.

ClassicalSpinMC.readMonteCarloParameters — Method

function readMonteCarloParameters(fn::H5)

Reads the MonteCarloParameters from a H5 object.

ClassicalSpinMC.readMonteCarloStatistics — Method

function readMonteCarloStatistics(fn::H5)

Reads the MonteCarloStatistics from a H5 object.

ClassicalSpinMC.readObservables — Method

function readObservables(fn::Union{HDF5.File,HDF5.Group})

Reads the observable type from a H5 object. It assumes that the appropriate constructor for the saved observable type exists. If no observable type is defined, it switches to the built-in observables.

ClassicalSpinMC.readUnitcell — Method

function readUnitcell(fn::H5)

Reads the unitcell from a H5 object.

ClassicalSpinMC.replicaExchange! — Method

function replicaExchange!(mc::MonteCarlo{T}, energy::Float64, betas::Vector{Float64}, channelsUp, channelsDown, label) where {T<:Lattice}

Performs a parallel replica exchange sweep based on parameters from the MonteCarlo object on the current worker.

ClassicalSpinMC.replicaExchange! — Method

function replicaExchange!(mc::MonteCarlo{T}, energy::Float64, betas::Vector{Float64}, channelsUp, channelsDown, label) where {T<:Lattice}

Performs a serial replica exchange sweep based on parameters from the MonteCarlo object on the current worker.

ClassicalSpinMC.resetBasis! — Method

function resetBasis!(unitcell::UnitCell{D}) where {D}

Resets the basis of a unitcell and all quantities that depend on the basis. This includes the interactions and interaction fields.

ClassicalSpinMC.rotateAboutAxis — Method

function rotateAboutAxis(spin::SVector{3,Float64}, axis::SVector{3,Float64}, angle::Float64)

Rotates a vector about an axis by a particular angle.

Assumptions on normalization

This assumes that the vector and axis are normalized for performance reasons.

ClassicalSpinMC.run! — Function

function run!(mcs::MonteCarloExchange, outfile::Union{String,Nothing}=nothing)

Dispatches run to perform a Replica Exchange (Parallel Tempering) run.

ClassicalSpinMC.run! — Method

function run!(mcs::MonteCarloAnnealing; outfile::Union{String,Nothing}=nothing)

Dispatches run to perform a Simulated Annealing run.

ClassicalSpinMC.run! — Method

function run!(mc::MonteCarlo{T}, betas::Vector{Float64}, channelsUp::Vector{RemoteChannel{Channel{C}}},
channelsDown::Vector{RemoteChannel{Channel{C}}},

Internal function used for parallel tempering runs.

ClassicalSpinMC.run! — Method

function run!(mc::MonteCarlo{T}; outfile::Union{String,Nothing}=nothing) where {T<:Lattice}

Dispatches run to perform a fixed-temperature Monte Carlo run.

ClassicalSpinMC.sanityChecks — Method

function sanityChecks(mc::MonteCarlo{T}, outfile::Union{String,Nothing}=nothing) where {T<:Lattice}

Perform sanity checks on a MonteCarlo struct to prevent undefined behaviour during the run.

ClassicalSpinMC.setField! — Method

function setField!(unitcell::UnitCell{D}, b::Int, B::SVector{3,Float64}) where {D}

Sets the magnetic field for a spin at the basis site 'b'.

ClassicalSpinMC.setField! — Method

function setField!(unitcell::UnitCell{D}, B::Vector{SVector{3,Float64}}) where {D}

Sets the magnetic field for all spins of the given unitcell.

ClassicalSpinMC.setInteractionOnsite! — Method

function setInteractionOnsite!(unitcell::UnitCell{D}, b::Int, M::SMatrix{3,3,Float64,9}) where {D}

Sets the self-interaction for a spin located at basis site 'b' of the given 'unitcell'. The exchange energy is calculated as spin'.M.spin

ClassicalSpinMC.setInteractionOnsite! — Method

function setInteractionOnsite!(unitcell::UnitCell{D}, M::Vector{SMatrix{3,3,Float64,9}}) where {D}

Sets the self-interaction for all spins of the given 'unitcell'.

The exchange energy for each spin is calculated as spins'[i].M[i].spins[i]

ClassicalSpinMC.setSpin! — Method

function setSpin!(lattice::Lattice{D,N}, site::Int, newState::SVector{3,Float64}) where {D,N}

Interface function to set the spin at a particular site.

ClassicalSpinMC.setSpin! — Method

function setSpin!(lattice::Lattice{D,N}, site::Int, newState::NTuple{3,Float64}) where {D,N}

Convenience function that converts a NTuple into SVector internally.

ClassicalSpinMC.sphericalUpdate — Function

function sphericalUpdate(rng=Random.GLOBAL_RNG)

Generates a random point on a unit sphere using cylindrical co-ordinates.

ClassicalSpinMC.updateTotalStatistics! — Method

function updateTotalStatistics!(statistics::MonteCarloStatistics)

Updates the running total of statistics while resetting statistics from the current sweep.

ClassicalSpinMC.writeCheckpoint! — Method

function writeCheckpoint!(filename::String, mc::MonteCarlo{Lattice{D,N}}) where {D,N}

Interface function to write a MonteCarlo checkpoint.

Checkpoints

Checkpoints use Serialization and thus are not recommended for long term storage as they may not be supported on a different Julia version.

ClassicalSpinMC.writeLattice! — Method

function writeLattice!(fn::H5, lattice::Lattice{D,N}) where {D,N}

Writes the Lattice in a H5 object.

ClassicalSpinMC.writeMonteCarlo! — Method

function writeMonteCarlo!(filename::String, mc::MonteCarlo)

Writes the MonteCarlo object into a HDF5 file.

ClassicalSpinMC.writeMonteCarlo! — Method

function writeMonteCarlo!(fn::H5, mc::MonteCarlo)

Writes the MonteCarlo object in a H5 object.

ClassicalSpinMC.writeMonteCarloParameters! — Method

function writeMonteCarloParameters!(fn::H5, mcp::MonteCarloParameters{U}) where {U}

Writes the MonteCarloParameters in a H5 object.

ClassicalSpinMC.writeMonteCarloStatistics! — Method

function writeMonteCarloStatistics!(fn::H5, mcs::MonteCarloStatistics)

Writes the MonteCarloStatistics in a H5 object.

ClassicalSpinMC.writeObservables! — Method

function writeObservables!(fn::Union{HDF5.File,HDF5.Group}, obs::Observables)

Writes the observable type in a H5 object. This is done by calling save! on each field.

ClassicalSpinMC.writeUnitcell! — Method

function writeUnitcell!(fn::H5, uc::UnitCell{D}) where {D}

Writes the unitcell in a H5 object.